workaround trusty init-system-helpers package version
If you install this package on a trusty distribution the following dependency would fail
init-system-helpers (>= 1.18~) on trusty only 1.14 is available.

Therefore Debian control can ignore the version adding a | in case of missing version
it will now ignore the hard dependency and just install what's available in the distribution repositories.